added unit tests for adding and retrieving simulation from database

+++ src/main/java/com/example/epidemicsimulator/domain/simulationcreator/DailyDataRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 interface DailyDataRepository extends JpaRepository<DailyData, Long> {
-    void deleteBySimulationId(int id);
+    void deleteBySimulationId(Long id);
 
     List<DailyData> findBySimulationId(Long id);
 }
